fix:iLogRequest.go 修改方法返回

+++ logexternal/logiface/iLogRequest.go
@@ -5,7 +5,7 @@ import (
 )
 
 type ILogRequest interface {
-	GetLogId() uint32           //日志随机号
+	GetLogId() uint64           //日志随机号
 	GetLogData() []byte         //获取日志请求消息的参数数据
 	GetLogLevel() logenum.Level //获取日志请求的级别
 	GetLogDataStr() string      //获取日志请求消息的参数数据转化成string
